updated the format of the select list, added site link to each tab and added labels
